Conditional Distance Covariance/Correlation Statistics

Computes conditional distance covariance and conditional distance correlation statistics, which are multivariate measures of conditional dependence.

cdcov(x, y, z, width, index = 1, distance = FALSE)

cdcor(x, y, z, width, index = 1, distance = FALSE)

Arguments

x

a numeric vector, matrix, or dist object

y

a numeric vector, matrix, or dist object

z

z is a numeric vector or matrix. It is the variable being conditioned.

width

a user-specified positive value (univariate conditional variable) or vector (multivariate conditional variable) for gaussian kernel bandwidth. Its default value is relies on stats::bw.nrd0.

index

exponent on Euclidean distance, in \((0,2]\)

distance

if distance = TRUE, x and y will be considered as distance matrices. Default: distance = FALSE.

Value

cdcov

conditional distance covariance test statistic.

cdcor

conditional distance correlation statistic.

cdc

conditional distance covariance/correlation vector.

Details

cdcov and cdcor compute conditional distance covariance and conditional distance correlation statistics. The sample sizes (number of rows or length of the vector) of the two variables must agree, and samples must not contain missing values. If we set distance = TRUE, arguments x, y can be a dist object recording distance between samples; otherwise, these arguments are treated as multivariate data.

References

Wang, X., Pan, W., Hu, W., Tian, Y. and Zhang, H., 2015. Conditional distance correlation. Journal of the American Statistical Association, 110(512), pp.1726-1734.

See also

Author

Canhong Wen, Wenliang Pan, and Xueqin Wang

Examples

library(cdcsis)

############# Conditional Distance Covariance #############
set.seed(1)
x <- rnorm(25)
y <- rnorm(25)
z <- rnorm(25)
cdcov(x, y, z)
#> $statistic
#>       cdcov 
#> 0.003188321 
#> 
#> $cdc
#>  [1] 0.05123402 0.05370986 0.05264026 0.05829258 0.04462193 0.17094674
#>  [7] 0.05649201 0.05625645 0.04659562 0.05830073 0.28539594 0.05823852
#> [13] 0.04324333 0.05785152 0.05308412 0.05580492 0.07790745 0.04789515
#> [19] 0.05638531 0.23012869 0.04919566 0.05314396 0.04544714 0.05431205
#> [25] 0.06185691
#> 
############# Conditional Distance Correlation #############
num <- 25
set.seed(1)
x <- rnorm(num)
y <- rnorm(num)
z <- rnorm(num)
cdcor(x, y, z)
#> $statistic
#>     cdcor 
#> 0.2391659 
#> 
#> $cdc
#>  [1] 0.1825805 0.1866731 0.1859597 0.2184754 0.1667906 0.4668055 0.1914210
#>  [8] 0.2084438 0.1706733 0.1966057 0.6438620 0.1971804 0.1614137 0.1967938
#> [15] 0.1880442 0.1930253 0.3234316 0.1770588 0.1942024 0.5637799 0.1774826
#> [22] 0.1873133 0.1675616 0.1980779 0.2354914
#>